pyproject.toml
[project]
name = "nhl-global-series-discord-bot"
version = "0.1.0"
description = "A Discord bot that notifies me as soon as tickets for the NHL Global Series games in Düsseldorf become available"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "discord-py>=2.7.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
 
[tool.hatch.build.targets.wheel]
packages = ["/nhl_global_series_discord_bot"]